// src/app/models/usuario.model.ts
export class Usuario { //creamos el usuario con los atributos requeridos
    nombres?:string;
    apellidos?: string;
    cc?: number;
    email?: string;
    fecha?: Date;
    direccion?: string;
    departamento?:string;
    ciudad?: string;
    pais?: string;
    codigo_postal?: number;
    profesion?: string[];
    habilidades?: string[];
    descripcion?: string;
    urlImagen?:string;
}